QualityMeasure update for nominal measures. ConvexHull.HullPoint coordinates are ALWAYS integers, so float replaced by int.
QualityMeasure:
Old calculate used float for the True Positive and Coverage counts.
This would cause rounding errors for large datasets.
Parameters are now int.
When required, these are converted to double, which is safe (no loss in precision).

Also, calculate now uses consistent naming for all measures, based on Head/Body.

Finally, all measures should now check for border cases of H or B==0, and NaNs.



ConvexHull.HullPoint:
HullPoint is a static inner class now, as it needs no reference to (members of)
the enclosing class.
The coordinates are integer counts (number of positives and negatives), so
instead of float they are now int.
This avoids the same issues for large datasets as QualityMeasure (see above).



ConvexHull:
See HullPoint float->int.
Internal calculations are adapted to not overflow.
Also, there should be less (no) rounding errors in the calculations.



SubgroupDiscovery + gui/ROCCurve are affected by all of the above.



git-svn-id: https://svn.science.uu.nl/repos/sci.knobb101.SubgroupDiscovery@2583 3a85a776-1598-11dc-ad04-adf9e166afdd

+++ SubDisc/src/nl/liacs/subdisc/gui/ROCCurve.java
@@ -118,8 +118,8 @@ public class ROCCurve extends JPanel
 		int N = itsQualityMeasure.getNrRecords();
 		int p = itsQualityMeasure.getNrPositives();
 		int aResolution = 400;
-		float aMax = Math.max(itsQualityMeasure.calculate(p, p), itsQualityMeasure.calculate(p, N));
-		aMax = Math.max(aMax, itsQualityMeasure.calculate(0, N-p));
+		float aMax = (float) Math.max(itsQualityMeasure.calculate(p, p), itsQualityMeasure.calculate(p, N));
+		aMax = (float) Math.max(aMax, itsQualityMeasure.calculate(0, N-p));
 		for (int i=0; i<aResolution; i++)
 		{
 			float anX = i/(float)aResolution;
@@ -128,7 +128,10 @@ public class ROCCurve extends JPanel
 			{
 				float aY = -(j+1)/(float)aResolution;
 				float aPositives = -aY*p; //this can be fractional, even though the counts are always integer
-				int aValue = (int) (255*itsQualityMeasure.calculate(aPositives, aNegatives+aPositives)/aMax);
+// FIXME MM
+// casting from fractional float to int might not yield best pictures
+// calculate(x, y) requires (x >= 0) and (y > 0), this needs to be guaranteed
+				int aValue = (int) (255 * itsQualityMeasure.calculate((int)aPositives, (int)(aNegatives+aPositives)) / aMax);
 				boolean isNegative = (aValue<0);
 				if (isNegative)
 					aValue = -aValue;
